refactor: файлики по папкам по-другому разложены
